Extract common logic for qualifying types
