+++ crates/analyzer/src/infer/rules/data.rs
@@ -1,9 +1,4 @@
-use crate::{
-    id::InFile,
-    infer::{ConstructorId, CoreTypeVariable},
-    surface::tree::*,
-    InferenceDatabase,
-};
+use crate::{id::InFile, infer::ConstructorId, surface::tree::*, InferenceDatabase};
 
 use super::{CoreType, InferContext};
 
@@ -26,29 +21,10 @@ impl InferContext<'_> {
 
         data_declaration.constructors.iter().for_each(|(constructor_id, data_constructor)| {
             let fields_ty = data_constructor.fields.iter().map(|field| self.lower_type(db, *field));
-
             let constructor_ty = fields_ty.rev().fold(data_ty, |data_ty, field_ty| {
                 db.intern_type(CoreType::Function(field_ty, data_ty))
             });
-
-            let qualified_ty = data_declaration.variables.iter().rev().fold(
-                constructor_ty,
-                |constructor_ty, variable| {
-                    let variable = match variable {
-                        TypeVariable::Kinded(name, kind) => {
-                            let name = Name::clone(name);
-                            let kind = self.lower_type(db, *kind);
-                            CoreTypeVariable::Kinded(name, kind)
-                        }
-                        TypeVariable::Name(name) => {
-                            let name = Name::clone(name);
-                            CoreTypeVariable::Name(name)
-                        }
-                    };
-                    db.intern_type(CoreType::Forall(variable, constructor_ty))
-                },
-            );
-
+            let qualified_ty = self.qualify_type(db, &data_declaration.variables, constructor_ty);
             self.state.map.of_constructor.insert(*constructor_id, qualified_ty);
         });
     }
